my_site/pyproject.toml
[tool.poetry]
name = "test"
version = "1"
description = "1"
authors = ["1"]
license = "1"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "<5.0.0"
django-rest-framework = "^0.1.0"
djangorestframework-simplejwt = "^5.3.1"
djangorestframework = "^3.14.0"
pyjwt = "^2.8.0"
simplejwt = "^2.0.1"
psycopg = "^3.1.18"
python-dotenv = "^1.0.1"
gunicorn = "^21.2.0"
whitenoise = "^6.6.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
